# fix(redhat): Also try to find buildinfo in root layer (layer 0)