@knqyf263 2025-05-27
This is what Red Hat shared for [the scanner certification](https://connect.redhat.com/en/partner-with-us/red-hat-vulnerability-scanner-certification). The first layer must not have any buildinfo. If RHEL 10 has buildinfo in the first layer, it's probably their mistake, or they've added a breaking change without notice.

![image](https://github.com/user-attachments/assets/c74287e0-19bc-49fd-b6d4-5a944403722f)