@Romain-Geissler-1A 2025-05-27
Independantly of Red Hat plans, as a RHEL client myself, I do deliver to folks in my company a RHEL base image, which is configured to target our own private mirror of RHEL rpm packages (so it can work without internet access). When I do that, I do build an OCI image which squashes all layers (Red Hat ones + the ones of my company to change the rpm configuration) into a single one. So if you consider this use case legit, this pull request will also "make it work". For me so far it was not seen with RHEL 9 as RHEL 9 had this concept of "default content set" which was actually hidding the problem. Technically the latest RHEL 9 image do have the same layout as the latest RHEL 10 images.