@DmitriyLewen 2025-05-27
hm... I didn't find diagram from https://github.com/aquasecurity/trivy/pull/8924#issuecomment-2911162458.

But i checked some images:
- registry.access.redhat.com/ubi7/ubi:latest  
- registry.access.redhat.com/ubi8/ubi:latest  
- registry.access.redhat.com/ubi9/ubi:latest  
- registry.access.redhat.com/ubi10/ubi:latest  
These images have only one layer, and that layer contains content sets.

Also i checked [test-harness images](https://redhat-connect.gitbook.io/partner-guide-red-hat-vulnerability-scanner-cert/workflow#certification-test-harness-images):
- registry.redhat.io/rhel9/python-312@sha256:a4b270f826f5ae08eefada3a73c989fd2cfb712df9d9b86cdce53ee201189a8c
- registry.redhat.io/openshift4/ose-console-rhel9@sha256:422c5c78b8fca878dba9d7faaae003e5341741b007fa51646904bc8d0c8a46c6
Both images have 4 layers. 0 layer also contains contents sets. 

So looks like we need to check `0` layer for these images.