@knqyf263 2025-05-28
Hmm... That differs from the explanation I received with the diagram. Nevertheless, if that is indeed the actual situation, then we will need to conform to it.

We also need to update the following logic. We may want to share the content set with the second layer only when the first layer doesn't have content set.
https://github.com/aquasecurity/trivy/blob/deefd67b3034a0dda5e8c87016d862579c88e7f9/pkg/fanal/applier/docker.go#L63-L70