@DmitriyLewen 2025-05-28
> We also need to update the following logic. We may want to share the content set with the second layer only when the first layer doesn't have content set.

I was thinking about this too and trying to find an image for this case.
I tested the image `registry.redhat.io/rhel9/rhel-bootc`.
This image contains many layers, but only the last layer contains content sets.
But there is one moment - all these layers don't have rpm packages.
So these changes don't make sense for this image.

Since we have confirmation (unlike the case when layer 0 contains set content) that there may be cases when we should take set content for layer 0 from >=layer 2 - I think it is worth sticking to the information you received from RedHat

the worst case that can happen is that we don't return the content for layer 0 (because layer 1 doesn't contain content sets either)
in this case, Trivy will take default content set instead of the content set from >=2 layers. 
But I'm not sure I understand how we can be sure that we should take the content set with >=2 layers. We reason this way only based on the information that we should take the content sets from layer 1, but as practice shows, this is not always correct.