@knqyf263 2025-05-28
>So on recent RHEL images, since layer 0 build info exists, we return it immediately, without trying to use layer 1 build info.

Ah, you're right. It already works as desired.
https://github.com/aquasecurity/trivy/blob/deefd67b3034a0dda5e8c87016d862579c88e7f9/pkg/fanal/applier/docker.go#L59-L61