@DmitriyLewen 2025-05-28
> I think it's acceptable to make this change while we wait for a response from Red Hat.

I don't mind